+++ it/typescript/.eslintrc.js
@@ -1,0 +1,45 @@
+module.exports = {
+  root: true,
+  env: {
+    es2022: true,
+    node: true,
+  },
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    project: ['tsconfig.json', 'tsconfig.eslint.json'],
+    sourceType: 'module',
+  },
+  extends: ['prettier'],
+  plugins: ['@typescript-eslint'],
+  rules: {
+    '@typescript-eslint/indent': ['error', 2],
+    '@typescript-eslint/member-delimiter-style': [
+      'error',
+      {
+        multiline: {
+          delimiter: 'semi',
+          requireLast: true,
+        },
+        singleline: {
+          delimiter: 'semi',
+          requireLast: false,
+        },
+      },
+    ],
+    '@typescript-eslint/quotes': [
+      'error',
+      'single',
+      {
+        avoidEscape: true,
+      },
+    ],
+    '@typescript-eslint/semi': ['error', 'always'],
+    'comma-dangle': ['error', 'always-multiline'],
+    'max-classes-per-file': 'off',
+    'no-console': 'error',
+    'no-multiple-empty-lines': ['error', { max: 1 }],
+    'no-redeclare': 'error',
+    'no-return-await': 'error',
+    'prefer-const': 'error',
+  },
+};