+++ it/typescript/src/typings/typings.d.ts
@@ -1,0 +1,2 @@
+declare const VERSION: string;
+declare const DEVELOP: boolean;