+++ src/Kiota.Builder/Lock/KiotaLockGenerationContext.cs
@@ -1,0 +1,8 @@
+﻿using System.Text.Json.Serialization;
+
+namespace Kiota.Builder.Lock;
+
+[JsonSerializable(typeof(KiotaLock))]
+internal partial class KiotaLockGenerationContext : JsonSerializerContext
+{
+}